Classifier/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opsera-classifier"
version = "0.1.0"
description = "Opsera KB similarity classifier with HITL loop"
requires-python = ">=3.10"
dependencies = [
  "chromadb>=0.5",
  "openai>=1.40",
  "pydantic>=2",
  "pydantic-settings>=2",
  "python-dotenv>=1.0",
  "langchain-text-splitters>=0.2",
  "langchain-openai>=0.1",
  "langchain-chroma>=0.1",
]

[project.scripts]
opsera-ingest = "opsera_classifier.ingest:main"
opsera-hitl   = "opsera_classifier.hitl:main"

[tool.setuptools.packages.find]
where = ["src"]
